Button Descriptions
Added button descriptions on WelcomePage and MapPage. On WelcomePage, the Get Orientation button is able to describe that the label above will be updated with the current phone orientation. On the map page, the get location button is now able to explain what will happen next (ask for permission and then show the current location).